Fix Year 2038 Problem in UserHistoryStorage
This CL fixes Year 2038 Problem only in UserHistoryStorage.  The same problem in other parts should be addressed by someone separately.  Note that it is guaranteed in proto binary format that replacing a uint32 entry with a uint64 item with keeping its field ID doesn't break binary compatibility.

BUG=none
TEST=unittest

git-svn-id: https://mozc.googlecode.com/svn/trunk@346 a6090854-d499-a067-5803-1114d4e51264

+++ src/prediction/user_history_predictor.h
@@ -445,7 +445,7 @@ class UserHistoryPredictor : public PredictorInterface {
       EntryPriorityQueue *results) const;
 
   void InsertHistory(bool is_suggestion_selected,
-                     uint32 last_access_time,
+                     uint64 last_access_time,
                      Segments *segments);
 
   // insert |key,value,description| to the internal dictionary database.
@@ -457,7 +457,7 @@ class UserHistoryPredictor : public PredictorInterface {
               const string &description,
               bool is_suggestion_selected,
               uint32 next_fp,
-              uint32 last_access_time,
+              uint64 last_access_time,
               Segments *segments);
 
   // Insert event entry (CLEAN_ALL_EVENT|CLEAN_UNUSED_EVENT).
